+++ test/bract/ring/inducer_test.clj
@@ -17,6 +17,43 @@
 (def holder (volatile! 10))
 
 
+(defn middleware-inc
+  [handler]
+  (vswap! holder (fn [^long n] (inc n)))
+  handler)
+
+
+(defn middleware-add
+  [handler ^long delta]
+  (vswap! holder (fn [^long n] (+ n delta)))
+  handler)
+
+
+(deftest test-apply-middleware
+  (testing "happy cases, wrapper fns"
+    (let [context {:bract.ring/ring-handler identity}]
+      (vreset! holder 0)
+      (let [new-context (inducer/apply-middlewares context [middleware-inc
+                                                            [middleware-add 2]])
+            new-handler (ring-kdef/ctx-ring-handler new-context)]
+        (is (contains? new-context :bract.ring/ring-handler))
+        (is (= :foo (new-handler :foo))))
+      (is (= 3 @holder))))
+  (testing "happy cases, wrapper names"
+    (let [context {:bract.ring/ring-handler identity}]
+      (vreset! holder 0)
+      (let [new-context (inducer/apply-middlewares context '[bract.ring.inducer-test/middleware-inc
+                                                             [bract.ring.inducer-test/middleware-add 2]])
+            new-handler (ring-kdef/ctx-ring-handler new-context)]
+        (is (contains? new-context :bract.ring/ring-handler))
+        (is (= :foo (new-handler :foo))))
+      (is (= 3 @holder))))
+  (testing "empty middlewares collection"
+    (inducer/apply-middlewares {} []))
+  (testing "missing Ring handler"
+    (is (thrown? IllegalArgumentException (inducer/apply-middlewares {} [(fn [x y] x)])))))
+
+
 (defn wrapper-inc
   [handler context]
   (vswap! holder (fn [^long n] (inc n)))